+++ CHANGES.md
@@ -4,6 +4,7 @@
 - CFR: Enhanced job statistics with optional reporting database support. Thanks, @WalBeh.
 - Settings: Added settings comparison utility. Thanks, @WalBeh.
 - Meta: Added parser for `https://cratedb.com/releases.json` file. Thanks, @WalBeh.
+- CFR: Added the ability to anonymize queries recorded by `collect`
 
 ## 2025/04/23 v0.0.32
 - MCP: Add subsystem providing a few server and client utilities through